p4tenant/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "p4tenant"
version = "0.1.0"
description = "CLI tool for managing P4-RESTART tenants"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "ruamel.yaml>=0.18.0",
    "pydantic>=2.0.0",
]

[project.scripts]
p4tenant = "p4tenant.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
